No conseguida comunicación de comandos bidireccional, readBytes no funciona

+++ Mini-bar/src/sensors/Sensor.c++
@@ -9,6 +9,6 @@ Sensor::Sensor(uint8_t pin,bool digital){
 
 //Valores entre 0 y 4095
 uint8_t Sensor::readValues(){
-    if(this->digital) return static_cast<uint16_t>(digitalRead(this->pin));
+    if(this->digital) return digitalRead(this->pin);
     else return analogRead(this->pin);
 }
